% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dokumentacja_dane_przykladowe.R
\docType{data}
\name{df_sample}
\alias{df_sample}
\title{Example data}
\format{
\subsection{\code{moj_df}}{

A data frame with 10 observations on 2 variables.
\describe{
\item{wiek}{aktualny wiek drzewostanu}
\item{y}{aktualna wysokość drzewostanu}
}
}
}
\source{
Manually created.
}
\usage{
df_sample
}
\description{
Dane przykładowe zawierajace infomracje o wieku i wysokośi drzewostanu.
}
\keyword{datasets}
