% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thmodel.R
\name{thmodel}
\alias{thmodel}
\title{Calculate TH for given age with actual age and height}
\usage{
thmodel(H1, T1, T2)
}
\arguments{
\item{H1}{Aktualna wysokość drzewostanu}

\item{T1}{Aktualny wiek drzewostanu}

\item{T2}{Wiek, dla którego liczymy wysokośc}
}
\value{
wektor numeryczny
}
\description{
Calculate TH for given age with actual age and height
}
\examples{
#' thmode(20, 70, 110)
}
